refactor: new folder structure for datasets
